Fix path handling when broxying to backend.
This was previously combining the backend path with the incoming request
path, resulting in a trailing / being added to some requests that wasn't
rpesent initially. This caused the route-service signature validation to
fail.

This includes refactoring things to avoid constructing a reverse proxy
for each request, and instead use a single instance. This can make it
more efficient as it can make use of things like HTTP keepalive etc.